requirements.txt
absl_py==1.4.0
click==8.1.4
numpy==1.25.1
pandas==2.0.3
python-dotenv==1.0.0
scikit_learn==1.3.0
setuptools==68.0.0
matplotlib==3.3.4
plotly==5.15.0
scipy==1.11.1
seaborn==0.11.1
statsmodels==0.12.2
